Delete all the code
The total amount of bugs is bug_density * lines_of_code, where bug
density is measured in bugs per line. So to make the total amount of
bugs go down, we can either decrease the bug density, or the lines of
code. To make the number of bugs go to zero, the easiest way is to just
delete all the code. </s>