Do not include the feeNumerator on the collector's state.
1. The user uses this endpoint to calculate the RewardsCollector address
   to pass to ChorusDeposit.deposit as withdrawalCredential before
deploying this RewardsCollector contract
2. The owner of FeeRewardsManager changes the defaultFeeNumerator

The user would not be able to deploy the RewardsCollector at the
predicted address since defaultFeeNumerator has been changed so the fees
get locked in that address unless the owner changes the
defaultFeeNumerator back to the value at step 1. so that the user would
be able to deploy the correct RewardsCollector.